Feat(DevTools): Unify Highlighter and TraceUpdates using a shared Canvas engine
Refactors the legacy DOM-injected element overlay system in Highlighter.js to paint directly onto a hardware-accelerated HTML5 Canvas, fulfilling a long-standing optimization request.

- Deprecates and removes legacy `Overlay.js` DOM node injection.
- Relocates `canvas.js` to a shared view directory to act as a global rendering hub.
- Implements `drawHighlighter` in `canvas.js` to render the translucent multi-layered CSS box model (Margin, Border, Padding, and Content bands) using native 2D context methods.
- Implements custom canvas-rendered tooltip boxes using `measureText` calculations to handle layout boundary clipping and collision avoidance natively.
- Streamlines `Highlighter.js` and `TraceUpdates/index.js` to pipe element metrics cleanly into the shared canvas entry configuration.

Closes #36787
